Use Balloons on Windows as notifications
They can be clicked to focus the window.

Add browser.js as preload, which includes wind32-browser.js if the app
is running on Windows. Since there doesn't seem to be a way to intercept
web notifications, we need to listen for XHR events and see if they're a
message. HipChat sends them as XML, since it's from the 1990's
apparently, so we need the xml2js node module.

I reallly hope Electron adds native support for notifications on Windows
soon... Follow their issue here:
https://github.com/atom/electron/issues/262