Add XO exception for Notification-shim/hack
We monkey patch Notification in browser-win32.js so we have
notifications in Windows. XO didn't like that ;) Adding an exception to
"no-native-reassign" for "Notification" helped, but we would still get
""Notification" is read only. (no-undef)" because it seems to think that
Notification is undefined when it recognized it in another rule. So it's
added to "globals" for now.