Simplify Toast notification on Windows
Make it more explicit by exposing the WindowsToaster part of
node-notifier. It has a fallback we should consider, where it'll use
balloons if no toast is available. It would simplify the settings and
code, and make it simpler for the user by having sensible defaults
(where they can't get both notification types at the sime time).